+++ bin/compilation_speed.ml
@@ -76,8 +76,6 @@ let benchmark_overhead backend () =
   in
   PrintBox_text.output Stdio.stdout plot_box;
   Stdio.print_endline "\n";
-  let module Backend = (val backend) in
-  Backend.unsafe_cleanup ();
   result
 
 let benchmarks =